requirements.txt
SQLAlchemy==1.3.22
plotly==5.5.0
matplotlib==3.5.1
pandas==1.3.5
statsmodels==0.13.1
psycopg2-binary==2.8.6
tenacity==6.3.1
multiprocessing-logging==0.3.1
seaborn==0.11.2
kaleido==0.2.1
scikit-learn==1.0.2
numpy==1.20.0
scikit-survival==0.17.0
dash==2.6.1
pandas_datareader==0.10.0
xgboost==1.2.1
census==0.8.19
shap==0.35.0